[JAVA_API] Load native libraries in static initializer blocks
* Add static initializer blocks to automatically load the required native libraries when the class is loaded

* Introduce the NativeLibrary class to encapsulate the logic for finding and loading native libs
