+++ modules/java_api/src/main/java/org/intel/openvino/Dimension.java
@@ -8,6 +8,15 @@ package org.intel.openvino;
  * shape-like object.
  */
 public class Dimension extends Wrapper {
+
+    static {
+        try {
+            Class.forName("org.intel.openvino.NativeLibrary");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException("Failed to load OpenVINO native libraries");
+        }
+    }
+
     protected Dimension(long addr) {
         super(addr);
     }
