+++ modules/java_api/src/main/java/org/intel/openvino/Output.java
@@ -6,6 +6,14 @@ package org.intel.openvino;
 /** A handle for one of a node's outputs. */
 public class Output extends Wrapper {
 
+    static {
+        try {
+            Class.forName("org.intel.openvino.NativeLibrary");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException("Failed to load OpenVINO native libraries");
+        }
+    }
+
     public Output(long addr) {
         super(addr);
     }
