+++ modules/java_api/src/main/java/org/intel/openvino/OutputInfo.java
@@ -17,6 +17,14 @@ package org.intel.openvino;
  */
 public class OutputInfo extends Wrapper {
 
+    static {
+        try {
+            Class.forName("org.intel.openvino.NativeLibrary");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException("Failed to load OpenVINO native libraries");
+        }
+    }
+
     public OutputInfo(long addr) {
         super(addr);
     }
