+++ modules/java_api/src/main/java/org/intel/openvino/PrePostProcessor.java
@@ -14,6 +14,14 @@ package org.intel.openvino;
  */
 public class PrePostProcessor extends Wrapper {
 
+    static {
+        try {
+            Class.forName("org.intel.openvino.NativeLibrary");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException("Failed to load OpenVINO native libraries");
+        }
+    }
+
     protected PrePostProcessor(long addr) {
         super(addr);
     }
