+++ modules/java_api/src/main/java/org/intel/openvino/PreProcessSteps.java
@@ -9,6 +9,14 @@ package org.intel.openvino;
  */
 public class PreProcessSteps extends Wrapper {
 
+    static {
+        try {
+            Class.forName("org.intel.openvino.NativeLibrary");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException("Failed to load OpenVINO native libraries");
+        }
+    }
+
     public PreProcessSteps(long addr) {
         super(addr);
     }
