+++ modules/java_api/src/main/java/org/intel/openvino/Wrapper.java
@@ -4,6 +4,15 @@
 package org.intel.openvino;
 
 public class Wrapper {
+
+    static {
+        try {
+            Class.forName("org.intel.openvino.NativeLibrary");
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException("Failed to load OpenVINO native libraries");
+        }
+    }
+
     protected final long nativeObj;
 
     protected Wrapper(long addr) {
